Refactor: validate 함수들을 일관성 있게 validator로 이동
